pridanie sk jazyka, pridané brány pre autorizovanie úkonov pre metódy update v EventControllery a destroy v AttendeeControllery

+++ event-management/app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
 use App\Http\Traits\CanLoadRelationships;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -60,8 +61,9 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
         return response(status: 204);
     }
